DHFPROD-4519: Always return static index file if resource not found
