--- explorer/src/main/java/com/marklogic/hub/explorer/WebMvcConfig.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2019 MarkLogic Corporation. All rights reserved.
- */
-package com.marklogic.hub.explorer;
-
-import java.io.IOException;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.resource.PathResourceResolver;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-  @Override
-  public void addResourceHandlers(ResourceHandlerRegistry registry) {
-
-    registry.addResourceHandler("/**/*")
-        .addResourceLocations("classpath:/static/")
-        .resourceChain(true)
-        .addResolver(new PathResourceResolver() {
-          @Override
-          protected Resource getResource(String resourcePath, Resource location)
-              throws IOException {
-            Resource requestedResource = location.createRelative(resourcePath);
-            return requestedResource.exists() && requestedResource.isReadable() ? requestedResource
-                : new ClassPathResource("/templates/index.html");
-          }
-        });
-  }
-}
